Replace perc params with hard coding
The percentage distribution of cpaital, wages, and income should not be
iterated. Rather it should allways sum to 100.
Hence, I have now defined them in config, and access them individually
in the simulation code.
1. User must only define 2 of the 3 values.
2. The third values is calculated.

The simulation code now individually sets the parameters in netlogo
model.
The excel data now stores the values.